Begin incorporating div_curl vector field functionality into the main library
